+++ src/index.html
@@ -27,7 +27,7 @@
           </a>
           <ul class="header__list">
             <li class="header__item">
-              <button class="header__btn curent js-home">Home</button>
+              <button class="header__btn current js-home">Home</button>
             </li>
             <li class="header__item">
               <button class="header__btn js-library">my library</button>
@@ -54,7 +54,7 @@
             </button>
           </form>
         </div>
-        <div class="header_buttons-conteiner display-none">
+        <div class="header_buttons-container display-none">
           <ul class="library-buttons-list">
             <li class="library-buttons-item">
               <button class="library-btn active js-watched">watched</button>
