+++ src/js/utils/refs.js
@@ -9,8 +9,8 @@ const refs = {
   libraryBtn: document.querySelector('.js-library'),
   watchedBtn: document.querySelector('.js-watched'),
   queueBtn: document.querySelector('.js-queue'),
-  libraryConteiner: document.querySelector('.header_buttons-conteiner'),
-  homeConteiner: document.querySelector('.header_form-container'),
+  libraryContainer: document.querySelector('.header_buttons-container'),
+  homeContainer: document.querySelector('.header_form-container'),
   paginationEl: document.querySelector('.footer__action'),
   headerEl: document.querySelector('header'),
   inputError: document.getElementById('input-error'),
