+++ src/sass/components/_header.scss
@@ -61,7 +61,7 @@
   
   color: #FFFFFF;
 
-  &.curent::after{
+  &.current::after{
     content: '';
     width: 85%;
     height: 3px;
@@ -152,7 +152,7 @@
   }
 }
 
-.header_buttons-conteiner{
+.header_buttons-container{
   display: flex;
   justify-content: center;
   margin-top: 40px;
@@ -174,7 +174,7 @@
 
 .library-btn{
   font-weight: 500;
-  font-style: 12px;
+  font-size: 12px;
   line-height: 1.33;
   color: #FFFFFF;
 
@@ -197,6 +197,7 @@
 .library-btn.active{
   background: #FF6B01;
   border: #FF6B01;
+  box-shadow: 0px 8px 43px rgba(255, 107, 1, 0.6);
 }
 
 .js-library-empty{
